[Auth] Update 'randomNonceString' implementation
