+++ authentication/LegacyAuthQuickstart/AuthenticationExampleSwift/MainViewController.swift
@@ -1041,41 +1041,26 @@ class MainViewController: UITableViewController {
   }
 
   // [START random_nonce]
-  // Adapted from https://auth0.com/docs/api-auth/tutorials/nonce#generate-a-cryptographically-random-nonce
   private func randomNonceString(length: Int = 32) -> String {
     precondition(length > 0)
-    let charset: [Character] =
-      Array("0123456789ABCDEFGHIJKLMNOPQRSTUVXYZabcdefghijklmnopqrstuvwxyz-._")
-    var result = ""
-    var remainingLength = length
-
-    while remainingLength > 0 {
-      let randoms: [UInt8] = (0 ..< 16).map { _ in
-        var random: UInt8 = 0
-        let errorCode = SecRandomCopyBytes(kSecRandomDefault, 1, &random)
-        if errorCode != errSecSuccess {
-          fatalError(
-            "Unable to generate nonce. SecRandomCopyBytes failed with OSStatus \(errorCode)"
-          )
-        }
-        return random
-      }
-
-      randoms.forEach { random in
-        if remainingLength == 0 {
-          return
-        }
-
-        if random < charset.count {
-          result.append(charset[Int(random)])
-          remainingLength -= 1
-        }
-      }
+    var randomBytes = [UInt8](repeating: 0, count: length)
+    let errorCode = SecRandomCopyBytes(kSecRandomDefault, randomBytes.count, &randomBytes)
+    if errorCode != errSecSuccess {
+      fatalError(
+        "Unable to generate nonce. SecRandomCopyBytes failed with OSStatus \(errorCode)"
+      )
     }
 
-    return result
-  }
+    let charset: [Character] =
+      Array("0123456789ABCDEFGHIJKLMNOPQRSTUVXYZabcdefghijklmnopqrstuvwxyz-._")
 
+    let nonce = randomBytes.map { byte in
+      // Pick a random character from the set, wrapping around if needed.
+      return charset[Int(byte) % charset.count]
+    }
+
+    return String(nonce)
+  }
   // [END random_nonce]
 
   // [START sha_256]
